Refactor project structure and add build scripts
