+++ .gitignore
@@ -1,0 +1,3 @@
+/build/
+u404shell
+*.o
